feat: enhance payment management with pagination and improved data handling
