+++ src/libs/Payment/getPayments.tsx
@@ -1,9 +1,12 @@
 import BACKEND_URL from "../config";
 
-export async function getPayments(token: string | undefined,  pageQuery?: string) {
-    let page = pageQuery ? `page=${pageQuery}` : '';
+export async function getPayments(token: string | undefined,  pageQuery?: string, statusQuery?: string) {
+    const queryParams = new URLSearchParams();
+    
+    if (pageQuery) queryParams.append("page", pageQuery);
+    if (statusQuery) queryParams.append("status", statusQuery);
 
-    const response = await fetch(`${BACKEND_URL}/api/v1/payments?${page}`, {
+    const response = await fetch(`${BACKEND_URL}/api/v1/payments?${queryParams.toString()}`, {
         method: "GET",
         headers: {
             Authorization: `Bearer ${token}`,
